# [Performance] Optimize CPU array creation in asarray for ~6× speedup (#27460)