@SwarnimShekhar 2025-03-26
Hello @yashk2810,
Thanks for pointing that out. I understand that wrapping `device_put` with `jit` doesn’t change its eager behaviour. My goal was to see if precompiling the call could reduce the overhead in repeated use cases, based on our benchmark results. It seems the benefit might not be as clear-cut, so I'm open to suggestions on alternative strategies or improvements.

Thanks again for your insights!