@jakevdp 2025-03-26
Hi - thanks for looking at this. I don't think we want this kind of specialization: it looks like it's about 50% faster than the current path, but the current solution does ignore some important things (e.g. `device` and `sharding`, dtype canonicalization, etc.). From a project maintenance standpoint, I don't think adding this extra complication for a marginal gain in one particular edge case is worth the cost.