@SwarnimShekhar 2025-03-26
Got it, that makes sense. I see how the trade-off isn’t worth it given the impact on device handling, `sharding`, and `dtype` canonicalization. Appreciate the feedback and the explanation!